+++ example/lib/platform/fox_player.dart
@@ -21,21 +21,21 @@ class FoxPlayer extends PlatformPlayer with HandleForces {
         );
 
   @override
-  void joystickChangeDirectional(JoystickDirectionalEvent event) {
+  void onJoystickChangeDirectional(JoystickDirectionalEvent event) {
     if (event.directional == JoystickMoveDirectional.MOVE_LEFT ||
         event.directional == JoystickMoveDirectional.MOVE_RIGHT ||
         event.directional == JoystickMoveDirectional.IDLE) {
-      super.joystickChangeDirectional(event);
+      super.onJoystickChangeDirectional(event);
     }
   }
 
   @override
-  void joystickAction(JoystickActionEvent event) {
+  void onJoystickAction(JoystickActionEvent event) {
     if (event.event == ActionEvent.DOWN &&
         (event.id == LogicalKeyboardKey.space.keyId || event.id == 1)) {
       jump();
     }
-    super.joystickAction(event);
+    super.onJoystickAction(event);
   }
 
   @override